fixes for table implementation.
Removed cell editing, because it is not the scope of the table system to
do someting like this. Got working the "Is Okaying" Column by correcting
the dataset and implementing a custom filter function. Also added some
graphical illustratition the boolean column.

Otherwise some formatting and unused import deletion happened as well.