migrated away from edge functions and deno - STARTING
This commit is not ment to be used in any test env, it serves only a
status report as there are other things to move before it works.

Introduced the `apps/api` nextjs project that will be the home for
serverless functions (hosted on Vercel, same place where we plan to host
our entire infra). This introduced Node.js in places where Deno APIs
were used, so I had to adjust our infra to node.

Because of this, Deno will be removed as a dependency and we will
migrate fully to nodejs. This involves migrating to node test
frameworks, removing deno configs and adjusting the readme, so expect
those in the following commits.