initial infra for packet detail formatting
Similarly to parsing, every implementation of `IPacketParser` will have
an implementation for `format`, a new method that transforms the given
`IPacketDetail` subclass into a matrix of strings that later will become
a table's body. There is also a new collector method
`formatPacketDetailsTable` that will generate the matrix based on it's
given input.

Also added the `implements IPacketParser` flag to
`DefaultStatusReportPacketParser`, because it looks like I missed it.

The reference was for the `IErrorPacketDetail` interface which also
covered in testing now.